Refactor plugin startup sequence starting only after 'init' hook
